// TypeScript-Client/src/Constants/enums/Stance.ts
// this may be incorrect or incomplete, but it's a start
enum Stance {
  airstrike = "airstrike",
  alert = "alert",
  alert2 = "alert2",
  alert3 = "alert3",
  alert4 = "alert4",
  alert5 = "alert5",
  alert6 = "alert6",
  assassination = "assassination",
  assassinationS = "assassinationS",
  assaulter = "assaulter",
  avenger = "avenger",
  backspin = "backspin",
  backstep = "backstep",
  blast = "blast",
  blizzard = "blizzard",
  brandish1 = "brandish1",
  brandish2 = "brandish2",
  burster1 = "burster1",
  burster2 = "burster2",
  cannon = "cannon",
  chainlightning = "chainlightning",
  coolingeffect = "coolineffect",
  darksight = "darksight",
  dash = "dash",
  dead = "dead",
  doublefire = "doublefire",
  doubleupper = "doubleupper",
  dragonstrike = "dragonstrike",
  eburster = "eburster",
  edrain = "edrain",
  eorb = "eorb",
  fake = "fake",
  fireburner = "fireburner",
  fist = "fist",
  fly = "fly",
  genesis = "genesis",
  ghostfly = "ghostfly",
  ghostjump = "ghostjump",
  ghostladder = "ghostladder",
  ghostproneStab = "ghostproneStab",
  ghostrope = "ghostrope",
  ghostsit = "ghostsit",
  ghoststand = "ghoststand",
  ghostwalk = "ghostwalk",
  handgun = "handgun",
  heal = "heal",
  holyshield = "holyshield",
  homing = "homing",
  info = "info",
  jump = "jump",
  ladder = "ladder",
  ladder2 = "ladder2",
  magic1 = "magic1",
  magic2 = "magic2",
  magic3 = "magic3",
  magic4 = "magic4",
  meteor = "meteor",
  ninjastorm = "ninjastorm",
  octopus = "octopus",
  paralyze = "paralyze",
  prone = "prone", // when click down
  prone2 = "prone2",
  proneStab = "proneStab",
  rapidfire = "rapidfire",
  recovery = "recovery",
  resurrection = "resurrection",
  rope = "rope",
  rope2 = "rope2",
  rush = "rush",
  rush2 = "rush2",
  sanctuary = "sanctuary",
  savage = "savage",
  screw = "screw",
  shoot1 = "shoot1",
  shoot2 = "shoot2",
  shoot6 = "shoot6",
  shootF = "shootF",
  shot = "shot",
  showdown = "showdown",
  sit = "sit",
  smokeshell = "smokeshell",
  somersault = "somersault",
  stabO1 = "stabO1",
  stabO2 = "stabO2",
  stabOF = "stabOF",
  stabT1 = "stabT1",
  stabT2 = "stabT2",
  stabTF = "stabTF",
  stand1 = "stand1",
  stand2 = "stand2",
  straight = "straight",
  swingO1 = "swingO1",
  swingO2 = "swingO2",
  swingO3 = "swingO3",
  swingOF = "swingOF",
  swingP1 = "swingP1",
  swingP2 = "swingP2",
  swingPF = "swingPF",
  swingT1 = "swingT1",
  swingT2 = "swingT2",
  swingT3 = "swingT3",
  swingTF = "swingTF",
  timeleap = "timeleap",
  torpedo = "torpedo",
  triplefire = "triplefire",
  walk1 = "walk1",
  walk2 = "walk2",
}

export enum MobStance {
  attack1 = "attack1",
  die1 = "die1",
  hit1 = "hit1",
  jump = "jump",
  move = "move",
  stand = "stand",
}

export default Stance;