// TypeScript-Client/src/Constants/enums/WZFiles.ts
enum WZFiles {
  Base = "Base.wz",
  Character = "Character.wz",
  Effect = "Effect.wz",
  Etc = "Etc.wz",
  Item = "Item.wz",
  Map = "Map.wz",
  Mob = "Mob.wz",
  Morph = "Morph.wz",
  Npc = "Npc.wz",
  Quest = "Quest.wz",
  Reactor = "Reactor.wz",
  Skill = "Skill.wz",
  Sound = "Sound.wz",
  String = "String.wz",
  TamingMob = "TamingMob.wz",
  UI = "UI.wz",
}

export default WZFiles;